Reorganize into a toolbox; add ngl shim.
The reorganization allows for multiple binaries, but for now there is still just the one.

The library adds a static shim wrapper for the NGL Library; this allows the rust layer to get at NGL's client data.  For now the only entry in the shim is a fetch of the machine ID, so we can use it to do activations/deactivations.