Set up to use git lfs with a free Artifactory server.
Because the NGL Library is too large for GitHub, we need to use git lfs to store it in a separate repository.  Because Adobe doesn't pay for GitHub LFS, and the Adobe internal Artifactory is not set up to support anonymous read access, they have advised me to get a free JFrog Cloud repository to hold the libraries.

Because git lfs requires local configuration to work properly, and because I don't think local configurations are stored in the repo, I'm copying the relevant lines from .git/config into this commit for reference:

[lfs]
	url = https://adobelfs1.jfrog.io/artifactory/api/lfs/libraries
[lfs "https://adobelfs1.jfrog.io/artifactory/api/lfs/libraries"]
	access = basic
	locksverify = false