Update to use new SSCP library.
The SSCP library has been revamped to be tiny, to return actual NGL base hardware IDs, and to do all its own compatibility testing with the NGL library.  So the toolbox now just uses it directly and doesn't bother doing compatibility testing.

This is still a mac-only build.  Link info for Windows to be verified.